+++ tests/test_os_open.py
@@ -40,3 +40,28 @@ class TestRemoteUrlToBrowserUrl:
     def testEmpty(self) -> None:
         assert remoteUrlToBrowserUrl("") is None
         assert remoteUrlToBrowserUrl("not-a-url") is None
+
+    # --------------------------------------------------------
+    # Method: testTreeBranch
+    # --------------------------------------------------------
+    def testTreeBranch(self) -> None:
+        assert (
+            remoteUrlToBrowserUrl("https://github.com/org/repo.git", branch="master")
+            == "https://github.com/org/repo/tree/master"
+        )
+        assert (
+            remoteUrlToBrowserUrl("git@github.com:org/repo.git", branch="main")
+            == "https://github.com/org/repo/tree/main"
+        )
+
+    # --------------------------------------------------------
+    # Method: testTreeBranchWithSlash
+    # --------------------------------------------------------
+    def testTreeBranchWithSlash(self) -> None:
+        assert (
+            remoteUrlToBrowserUrl(
+                "https://github.com/org/repo.git",
+                branch="feature/foo",
+            )
+            == "https://github.com/org/repo/tree/feature/foo"
+        )